Build features addTask & displayTaskList
